posix: timer: use explicit cast for relative duration conversion
ts_to_ms() returns int64_t but the result was assigned to a uint32_t
variable without an explicit cast, relying on implicit narrowing.
Add an explicit (uint32_t) cast to document the intentional truncation
for relative timeouts, which are expected to be small values that fit
within uint32_t.

+++ subsys/portability/posix/options/timer.c
@@ -275,7 +275,7 @@ int timer_settime(timer_t timerid, int flags, const struct itimerspec *value,
 	if ((flags & TIMER_ABSTIME) != 0) {
 		duration = timespec_to_timeoutms(timer->clock_id, &value->it_value);
 	} else {
-		duration = ts_to_ms(&value->it_value);
+		duration = (uint32_t)ts_to_ms(&value->it_value);
 	}
 
 	if (timer->status == ACTIVE) {
